requirement.txt
Appium-Python-Client
-U pytest